Roll back to JetBrains Platform 2024
We frequently had an issue opening the diagrams.net canvas with the new version.
It was somehow flaky (worked sometimes, but not stable) and even seems to have
caused a general problem with JCEF (we had to stop all JetBrains instances,
clean all JCEF-Caches, and then could restart (having it working again).